remove duplicate method assertions
